preload/modulepreload-as.html

<!doctype html>
<meta charset=utf-8>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  setup({ explicit_done: true });

  // Compared to modulepreload.html, this tests behavior when elements are
  // initially on an HTML page instead of being added by Javascript.
  const scriptLikes = [
    'audioworklet',
    'paintworklet',
    'script',
    'serviceworker',
    'sharedworker',
    'worker',
  ];

  const goodAsValues = ['', 'invalid-dest', 'sCrIpT', 'style', 'json',  ...scriptLikes];
  let testCount = 0;
  let all_links = document.querySelectorAll('link');

  // `all_links` may change once the DOM finishes loading.
  document.addEventListener('DOMContentLoaded', function() {
    all_links = document.querySelectorAll('link');
  }, false);

  function checkDone() {
    if (testCount === (all_links.length)) {
      done();
    }
  }

  function handleLoad(target) {
    const asValue = target.dataset.as; // Don't depend on "as" attribute reflection.
    const good = goodAsValues.includes(asValue);
    const preload_url = new URL(target.href, location.href);

    test(function() {
      assert_true(good);
    }, `Modulepreload expected load with as="${asValue}"`);

    test(function() {
      const downloads = performance
              .getEntriesByName(preload_url)
              .length;
            assert_equals(downloads, good ? 1 : 0);
    }, `Modulepreload loaded with as="${asValue}"`);

    ++testCount;
    checkDone();
  }

  function handleError(target) {
    const asValue = target.dataset.as; // Don't depend on "as" attribute reflection.
    const good = goodAsValues.includes(asValue);
    const preload_url = new URL(target.href, location.href);

    test(function() {
      assert_false(good);
    }, `Modulepreload expected error with as="${asValue}"`);

    test(function() {
      const downloads = performance
              .getEntriesByName(preload_url)
              .length;
            assert_equals(downloads, good ? 1 : 0);
    }, `Modulepreload expected no file loaded with error when as="${asValue}"`);

    ++testCount;
    checkDone();
  }
</script>
<link rel="modulepreload" href="resources/module1.js?empty-string" as="" data-as="" onload="handleLoad(this)" onerror="handleError(this)">
<link rel="modulepreload" href="resources/module1.js?audio" as="audio" data-as="audio" onload="handleLoad(this)" onerror="handleError(this)">
<link rel="modulepreload" href="resources/module1.js?audioworklet" as="audioworklet" data-as="audioworklet" onload="handleLoad(this)" onerror="handleError(this)">
<link rel="modulepreload" href="resources/module1.js?document" as="document" data-as="document" onload="handleLoad(this)" onerror="handleError(this)">
<link rel="modulepreload" href="resources/module1.js?embed" as="embed" data-as="embed" onload="handleLoad(this)" onerror="handleError(this)">
<link rel="modulepreload" href="resources/module1.js?font" as="font" data-as="font" onload="handleLoad(this)" onerror="handleError(this)">
<link rel="modulepreload" href="resources/module1.js?frame" as="frame" data-as="frame" onload="handleLoad(this)" onerror="handleError(this)">
<link rel="modulepreload" href="resources/module1.js?iframe" as="iframe" data-as="iframe" onload="handleLoad(this)" onerror="handleError(this)">
<link rel="modulepreload" href="resources/module1.js?image" as="image" data-as="image" onload="handleLoad(this)" onerror="handleError(this)">
<link rel="modulepreload" href="resources/module1.js?manifest" as="manifest" data-as="manifest" onload="handleLoad(this)" onerror="handleError(this)">
<link rel="modulepreload" href="resources/module1.js?object" as="object" data-as="object" onload="handleLoad(this)" onerror="handleError(this)">
<link rel="modulepreload" href="resources/module1.js?paintworklet" as="paintworklet" data-as="paintworklet" onload="handleLoad(this)" onerror="handleError(this)">
<link rel="modulepreload" href="resources/module1.js?report" as="report" data-as="report" onload="handleLoad(this)" onerror="handleError(this)">
<link rel="modulepreload" href="resources/module1.js?script" as="script" data-as="script" onload="handleLoad(this)" onerror="handleError(this)">
<link rel="modulepreload" href="resources/module1.js?serviceworker" as="serviceworker" data-as="serviceworker" onload="handleLoad(this)" onerror="handleError(this)">
<link rel="modulepreload" href="resources/module1.js?sharedworker" as="sharedworker" data-as="sharedworker" onload="handleLoad(this)" onerror="handleError(this)">
<link rel="modulepreload" href="resources/module1.js?track" as="track" data-as="track" onload="handleLoad(this)" onerror="handleError(this)">
<link rel="modulepreload" href="resources/module1.js?video" as="video" data-as="video" onload="handleLoad(this)" onerror="handleError(this)">
<link rel="modulepreload" href="resources/module1.js?webidentity" as="webidentity" data-as="webidentity" onload="handleLoad(this)" onerror="handleError(this)">
<link rel="modulepreload" href="resources/module1.js?worker" as="worker" data-as="worker" onload="handleLoad(this)" onerror="handleError(this)">
<link rel="modulepreload" href="resources/module1.js?xslt" as="xslt" data-as="xslt" onload="handleLoad(this)" onerror="handleError(this)">
<link rel="modulepreload" href="resources/module1.js?fetch" as="fetch" data-as="fetch" onload="handleLoad(this)" onerror="handleError(this)">
<link rel="modulepreload" href="resources/module1.js?invalid-dest" as="invalid-dest" data-as="invalid-dest" onload="handleLoad(this)" onerror="handleError(this)">
<link rel="modulepreload" href="resources/module1.js?iMaGe" as="iMaGe" data-as="iMaGe" onload="handleLoad(this)" onerror="handleError(this)">
<link rel="modulepreload" href="resources/module1.js?sCrIpT" as="sCrIpT" data-as="sCrIpT" onload="handleLoad(this)" onerror="handleError(this)">
<link rel="modulepreload" href="resources/dummy.css?style" as="style" data-as="style" onload="handleLoad(this)" onerror="handleError(this)">
<link rel="modulepreload" href="resources/dummy.json?json" as="json" data-as="json" onload="handleLoad(this)" onerror="handleError(this)">
<body>
